requirements.txt
matplotlib~=3.6.2
numpy~=1.23.0
torch~=2.1.1
setuptools~=49.2.1
numpy-stl~=2.17.0
pdb2pqr~=3.6.2